requirement.txt
numpy
pandas
yfinance
seaborn
matplotlib.pyplot